Flytt logikk tilhørende Periodeklassen for seg selv
