+++ src/main/kotlin/no/nav/familie/ks/sak/common/tidslinje/Periode.kt
@@ -1,0 +1,14 @@
+package no.nav.familie.ks.sak.common.tidslinje
+
+import java.time.LocalDate
+
+data class Periode<T>(
+    val verdi: T?,
+    val fom: LocalDate?,
+    val tom: LocalDate?
+) {
+
+    fun tilTidslinjePeriodeMedDato() = TidslinjePeriodeMedDato(verdi, fom, tom)
+}
+
+fun <T> List<Periode<T>>.tilTidslinje(): Tidslinje<T> = this.map { it.tilTidslinjePeriodeMedDato() }.tilTidslinje()
