+++ src/main/kotlin/no/nav/familie/ks/sak/common/tidslinje/TidslinjePeriodeMedDato.kt
@@ -3,15 +3,6 @@ package no.nav.familie.ks.sak.common.tidslinje
 import no.nav.familie.ks.sak.common.exception.Feil
 import java.time.LocalDate
 
-data class Periode<T>(
-    val verdi: T?,
-    val fom: LocalDate?,
-    val tom: LocalDate?
-) {
-
-    fun tilTidslinjePeriodeMedDato() = TidslinjePeriodeMedDato(verdi, fom, tom)
-}
-
 data class TidslinjePeriodeMedDato<T>(
     val periodeVerdi: PeriodeVerdi<T>,
     val fom: Dato,
@@ -45,9 +36,6 @@ data class TidslinjePeriodeMedDato<T>(
 
     fun tilPeriode() = Periode(periodeVerdi.verdi, fom.tilLocalDateEllerNull(), tom.tilLocalDateEllerNull())
 }
-
-@JvmName("perioderTilTidslinje")
-fun <T> List<Periode<T>>.tilTidslinje(): Tidslinje<T> = this.map { it.tilTidslinjePeriodeMedDato() }.tilTidslinje()
 
 fun <T> List<TidslinjePeriodeMedDato<T>>.tilTidslinje(): Tidslinje<T> {
     val perioder = this.tilTidslinjePerioder()
